rename BlankSlateView component to NoRepositoriesView
also rename css id and files
