--- app/src/ui/blank-slate/index.ts
@@ -1,1 +1,0 @@
-export { BlankSlateView } from './blank-slate'
