+++ app/src/ui/no-repositories/index.ts
@@ -1,0 +1,1 @@
+export { NoRepositoriesView } from './no-repositories-view'
