+++ app/styles/_ui.scss
@@ -53,7 +53,7 @@
 @import 'ui/call-to-action';
 @import 'ui/acknowledgements';
 @import 'ui/vertical-segmented-control';
-@import 'ui/blank-slate';
+@import 'ui/no-repositories';
 @import 'ui/terms-and-conditions';
 @import 'ui/ref';
 @import 'ui/monospaced';
